+++ src/components/navigation-bar/navigation-bar.jsx
@@ -13,6 +13,7 @@ export const NavigationBar = ({ user, onLoggedOut }) => {
                             <Navbar.Toggle aria-controls="basic-navbar-nav" />
                             <Navbar.Collapse id="basic-navbar-nav">
                                 <Nav className="me-auto">
+                                    <Nav.Link style={{ color: "white" }} href="/">Movies</Nav.Link>
 
                                     <Nav.Link style={{ color: "white" }} href="/profile">Profile</Nav.Link>
                                 </Nav>
